requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
defusedxml==0.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.7
django-allauth==0.39.1
django-filter==2.1.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
djangorestframework-stubs==0.3.0
gunicorn==19.9.0
idna==2.8
isort==4.3.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.8
pylint==2.3.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
static3==0.7.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
